+++ neuro_extras/seldon.py
@@ -1,0 +1,97 @@
+from distutils import dir_util
+from pathlib import Path
+from typing import Any, Dict
+
+import click
+from neuromation import api as neuro_api
+from neuromation.api.url_utils import uri_from_cli
+from yarl import URL
+
+from neuro_extras.data_copier import NEURO_EXTRAS_IMAGE
+from neuro_extras.image_builder import ImageBuilder
+
+
+ASSETS_PATH = Path(__file__).resolve().parent / "assets"
+SELDON_CUSTOM_PATH = ASSETS_PATH / "seldon.package"
+
+
+async def _init_seldon_package(path: str) -> None:
+    async with neuro_api.get() as client:
+        uri = uri_from_cli(
+            path,
+            client.username,
+            client.cluster_name,
+            allowed_schemes=("file", "storage"),
+        )
+        click.echo(f"Copying a Seldon package scaffolding into {uri}")
+        if uri.scheme == "file":
+            dir_util.copy_tree(str(SELDON_CUSTOM_PATH), path)
+        else:
+            await client.storage.mkdir(uri, parents=True)
+            await client.storage.upload_dir(URL(SELDON_CUSTOM_PATH.as_uri()), uri)
+
+
+async def _create_seldon_deployment(
+    *,
+    name: str,
+    neuro_secret_name: str,
+    registry_secret_name: str,
+    model_image_uri: str,
+    model_storage_uri: str,
+) -> Dict[str, Any]:
+    async with neuro_api.get() as client:
+        builder = ImageBuilder(client)
+        model_image_ref = builder.parse_image_ref(model_image_uri)
+
+    pod_spec = {
+        "volumes": [
+            {"emptyDir": {}, "name": "neuro-storage"},
+            {"name": "neuro-secret", "secret": {"secretName": neuro_secret_name}},
+        ],
+        "imagePullSecrets": [{"name": registry_secret_name}],
+        "initContainers": [
+            {
+                "name": "neuro-download",
+                "image": NEURO_EXTRAS_IMAGE,
+                "imagePullPolicy": "Always",
+                "command": ["bash", "-c"],
+                "args": [
+                    "cp -L -r /var/run/neuro/config /root/.neuro;"
+                    "chmod 0700 /root/.neuro;"
+                    "chmod 0600 /root/.neuro/db;"
+                    f"neuro cp {model_storage_uri} /storage"
+                ],
+                "volumeMounts": [
+                    {"mountPath": "/storage", "name": "neuro-storage"},
+                    {"mountPath": "/var/run/neuro/config", "name": "neuro-secret"},
+                ],
+            }
+        ],
+        "containers": [
+            {
+                "name": "model",
+                "image": model_image_ref,
+                "imagePullPolicy": "Always",
+                "volumeMounts": [{"mountPath": "/storage", "name": "neuro-storage"}],
+            }
+        ],
+    }
+    return {
+        "apiVersion": "machinelearning.seldon.io/v1",
+        "kind": "SeldonDeployment",
+        "metadata": {"name": name},
+        "spec": {
+            "predictors": [
+                {
+                    "componentSpecs": [{"spec": pod_spec}],
+                    "graph": {
+                        "endpoint": {"type": "REST"},
+                        "name": "model",
+                        "type": "MODEL",
+                    },
+                    "name": "predictor",
+                    "replicas": 1,
+                }
+            ]
+        },
+    }
